Add files via upload
This makes it so that pressing the ` key results in ESC being pressed while pressing SHIFT + ` key results in ~ being pressed.
Also. former INSERT key has been set to DELETE, while SHIFT + INSERT has been set to INSERT.

This is so that the first iteration of the PCB layout for the keyboard can become more useful in the current situation where the module is not working.